backend/analysis_service/requirements.txt
flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0
torch>=2.2.0
transformers>=4.42.0
numpy>=1.26.0
python-dotenv==1.0.0
xgboost>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
